Move the embed code in misc to a separate file
